requirements.txt
annotated-types==0.7.0
anyio==4.11.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
google-auth==2.43.0
google-genai==1.50.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
